+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 Unreleased
 ----------
 
+* Add CRFDI categorization and scripts to generate it from the UNFCCC DI flexible
+  query API.
+
 
 0.5.0 (2021-03-23)
 ------------------